requirements.txt
# --extra-index-url https://pypi.tuna.tsinghua.edu.cn/simple

accelerate
annotated-types==0.7.0
anyio==4.9.0
av==14.4.0
certifi==2025.6.15
charset-normalizer==3.4.2
deepspeed==0.17.1
packaging==25.0
distro==1.9.0
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.18.0
# flash_attn==2.8.0.post2
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.4
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.93.0
opencv-python==4.11.0.86
peft==0.15.2
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2024.11.6
requests
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn
scipy==1.15.3
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile
timm
tokenizers==0.21.1
tqdm==4.67.1
transformers
triton
typing-inspection
typing_extensions
urllib3
wandb
# modelscope
sentencepiece
rouge-score
bert-score
nltk
sentence-transformers
nvitop